requirements.txt
GeoAlchemy2~=0.15.0
SQLAlchemy~=2.0.29
uvicorn~=0.29.0
fastapi~=0.110.2
geojson~=3.1.0
pydantic~=2.7.1
python-dotenv~=1.0.1
pydantic_settings
asyncpg